remove conversion warnings

+++ src/JPEG12.cpp
@@ -255,9 +255,9 @@ static size_t nzeros(void* src, jpeg_params& params)
 {
     auto s = reinterpret_cast<JSAMPLE*>(src);
     size_t nzeros = 0;
-    int w = params.raster.size.x;
-    int h = params.raster.size.y;
-    int bands = params.raster.size.c;
+    int w = (int)params.raster.size.x;
+    int h = (int)params.raster.size.y;
+    int bands = (int)params.raster.size.c;
     for (int y = 0; y < h; y++) {
         for (int x = 0; x < w; x++) {
             bool zero = true;
@@ -278,7 +278,7 @@ static size_t update_mask(BitMask& mask, void* src, jpeg_params& params)
     size_t nzeros = 0;
     int w = mask.getWidth();
     int h = mask.getHeight();
-    int bands = params.raster.size.c;
+    int bands = (int)params.raster.size.c;
     for (int y = 0; y < h; y++) {
         for (int x = 0; x < w; x++) {
             bool zero = true;
@@ -324,7 +324,7 @@ const char *jpeg12_encode(jpeg_params &params, storage_manager &src, storage_man
 
     // might need a mask
     if (nzeros(src.buffer, params) > 0) {
-        BitMask mask(rsize.x, rsize.y);
+        BitMask mask((unsigned int)rsize.x, (unsigned int)rsize.y);
         update_mask(mask, src.buffer, params);
         jh.zenChunk.size = 2 * mask.size();
         maskbuff.resize(jh.zenChunk.size + CHUNK_NAME_SIZE);
@@ -361,7 +361,7 @@ const char *jpeg12_encode(jpeg_params &params, storage_manager &src, storage_man
 
     jpeg_start_compress(&cinfo, TRUE);
     jpeg_write_marker(&cinfo, JPEG_APP0 + 3, 
-        (JOCTET *)jh.zenChunk.buffer, jh.zenChunk.size);
+        (JOCTET *)jh.zenChunk.buffer, (unsigned int)jh.zenChunk.size);
 
     const JSAMPROW rowbuffer = reinterpret_cast<JSAMPROW>(src.buffer);
     while (cinfo.next_scanline != cinfo.image_height) {
